% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Data in the WNAI dataset.}
\format{A data frame with 172 societies (rows) and 430 variables (columns):
\describe{
\item{\code{soc_id}}{D-PLACE-wide id for the society}
\item{\code{WNAI8..WNAI438}}{429 variables for each society}
}}
\source{
Jorgensen, J. G. (1980) \emph{Western Indians: Comparative Environments,
Languages and Cultures of 172 Western American Indian Tribes}. W. H.
Freeman.

Jorgensen, J. G. (1999) Codebook for Western Indians data. \emph{World Cultures}
10(2): 144-293.

The dataset in D-PLACE is imported from \emph{World Cultures} CD-ROM (2011)
using codes described in Jorgensen (1999).
}
\usage{
data
}
\description{
Data in the WNAI dataset.
}
\keyword{datasets}
