% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.R
\docType{data}
\name{dplace_rev}
\alias{dplace_rev}
\title{D-PLACE revision of the WNAI dataset.}
\format{An object of class \code{character} of length 1.}
\usage{
dplace_rev
}
\value{
the D-PLACE revision
}
\description{
D-PLACE revision of the WNAI dataset.
}
\keyword{datasets}
