% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codes}
\alias{codes}
\title{Codes for each variable in the WNAI dataset.}
\format{A nested data frame with 429 pairs of variable id and codes (rows):
\describe{
\item{\code{var_id}}{D-PLACE-wide id for the variable}
\item{\code{codes}}{codes used in the variable:
\describe{
\item{\code{code}}{code in the variable}
\item{\code{name}}{short description of the code}
\item{\code{description}}{longer description of the code}
}}
}}
\source{
Jorgensen, J. G. (1980) \emph{Western Indians: Comparative Environments,
Languages and Cultures of 172 Western American Indian Tribes}. W. H.
Freeman.

Jorgensen, J. G. (1999) Codebook for Western Indians data. \emph{World Cultures}
10(2): 144-293.

The dataset in D-PLACE is imported from \emph{World Cultures} CD-ROM (2011)
using codes described in Jorgensen (1999).
}
\usage{
wnai::codes
}
\description{
Codes for each variable in the WNAI dataset.
}
\keyword{datasets}
