% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{societies}
\alias{societies}
\title{Societies in the WNAI dataset.}
\format{A data frame with 172 societies (rows) and 6 variables (columns):
\describe{
\item{\code{id}}{D-PLACE-wide id for the society}
\item{\code{name}}{name of the society}
\item{\code{alt_names}}{alternative names of the society (comma separated)}
\item{\code{glottocode}}{code of the most specific Glottolog languoid for
the society}
\item{\code{latitude}}{location of the society in latitude}
\item{\code{longitude}}{location of the society in longitude}
}}
\source{
Jorgensen, J. G. (1980) \emph{Western Indians: Comparative Environments,
Languages and Cultures of 172 Western American Indian Tribes}. W. H.
Freeman.

Jorgensen, J. G. (1999) Codebook for Western Indians data. \emph{World Cultures}
10(2): 144-293.

The dataset in D-PLACE is imported from \emph{World Cultures} CD-ROM (2011)
using codes described in Jorgensen (1999).
}
\usage{
wnai::societies
}
\description{
Societies in the WNAI dataset.
}
\keyword{datasets}
